checking if it exists before delet

+++ apps/server/src/services/RestoreService.ts
@@ -111,7 +111,9 @@ class RestoreService {
     }
 
     clear() {
-        fs.unlinkSync(path.join(this.filePath, 'restore.csv'));
+        if (fs.existsSync(path.join(this.filePath, 'restore.csv'))) {
+            fs.unlinkSync(path.join(this.filePath, 'restore.csv'));
+        }
     }
 }
